refactor: rename file_reader.ts to file-reader.ts for consistency
refactor: update import paths for FileReader to use hyphenated naming convention
